Committing lots of broken code
The Amazon API is proving to be a disaster. I'm not trying to pay for
the service, and I can't seem to get all of the libraries to work just
right. I'm either having .NET issues or I just can't seem to find the
solution online. So that's been fun.